remove: all 다 지우고 다시 시작
