+++ src/main/java/Application.java
@@ -1,8 +1,6 @@
-import controller.BlackjackController;
-
 public class Application {
     public static void main(String[] args) {
-        BlackjackController blackjackController = new BlackjackController();
-        blackjackController.run();
+        BlackjackStarter blackjackStarter = new BlackjackStarter();
+        blackjackStarter.run();
     }
 }
