+++ src/main/java/BlackjackStarter.java
@@ -1,0 +1,5 @@
+public class BlackjackStarter {
+
+    public void run() {
+    }
+}
