+++ src/main/java/domain/GameRoom.java
@@ -1,0 +1,5 @@
+package domain;
+
+public class GameRoom {
+
+}
