+++ src/main/java/domain/Profit.java
@@ -1,0 +1,10 @@
+package domain;
+
+public class Profit {
+    
+    private final int value;
+
+    public Profit(int value) {
+        this.value = value;
+    }
+}
