+++ src/main/java/domain/card/Card.java
@@ -1,35 +1,11 @@
 package domain.card;
 
-import java.util.List;
-
-public class Card {
-    private final Suit suit;
+public abstract class Card {
     private final Rank rank;
-    private boolean isOpened;
+    private final Suit suit;
 
-    public Card(Suit suit, Rank rank) {
-        this.suit = suit;
+    public Card(Rank rank, Suit suit) {
         this.rank = rank;
-        this.isOpened = false;
-    }
-
-    public void open() {
-        this.isOpened = true;
-    }
-
-    public List<Integer> getScores() {
-        return rank.getScores();
-    }
-
-    public Rank getRank() {
-        return rank;
-    }
-
-    public Suit getSuit() {
-        return suit;
-    }
-
-    public boolean isOpened() {
-        return isOpened;
+        this.suit = suit;
     }
 }
