--- src/main/java/domain/profit/ProfitStrategy.java
@@ -1,8 +1,0 @@
-package domain.profit;
-
-import domain.BattleResult;
-import domain.Bet;
-
-public interface ProfitStrategy {
-    Profit calculateProfit(Bet bet, BattleResult battleResult);
-}
